% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_midponits.R
\name{get_midpoints}
\alias{get_midpoints}
\title{Calculate Midpoints}
\usage{
get_midpoints(bin_factor)
}
\description{
Computes midpoints of suitability index (SI) bins.
}
\details{
use in the plot_SI_curve function for plotting
}
