% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSImap.R
\name{HSImap}
\alias{HSImap}
\title{Plot HSI Map}
\usage{
HSImap(surveydata, MapData, envVariable, Seasons, int_n, weights, filename)
}
\arguments{
\item{filename}{save the plotted HSI map with defined name.}
}
\description{
Creates spatial maps of habitat suitability indices.
}
\examples{
HSImap=HSImap(surveydata = Lobster_SI_9020,
MapData = LobsterHSI_mapData,
envVariable = c("Latitude","Longitude","Depth", "SST"),
Seasons =c("Spring","Fall"),
int_n = 20,
weights=c(0.25,0.25,0.25,0.25),
filename="HSIplot.jpg")

for project users, define months 1-6 as spring, define months 7-12 as fall
LobsterHSI_mapData <- LobsterHSI_monthData \%>\%
   dplyr::mutate(Season = case_when(
     Month \%in\% c(1:6) ~ "Spring",
     Month \%in\% c(7:12)   ~ "Fall",
     TRUE               ~ NA_character_))\%>\%
   dplyr::select(-Month)
}
