% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Lobster_SI_9020}
\alias{Lobster_SI_9020}
\title{Lobster Survey Data (1990–2020)}
\format{
A dataframe with rows = 13285, columns = 7:
\describe{
  \item{Year}{Numeric, sample year.}
  \item{Season}{Numeric, sample season.}
  \item{Latitude}{Numeric, station latitude.}
  \item{Longitude}{Numeric, station longitude.}
  \item{Depth}{Numeric, depth in meters.}
  \item{SST}{Numeric, sea surface temperature (°C).}
  \item{abundance}{Numeric, adult lobster abundance.}
}
}
\source{
Several Lobster Surveys
}
\usage{
Lobster_SI_9020
}
\description{
Biological survey data of lobsters used in SI curve generation.
}
\keyword{datasets}
