% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LobsterHSI_mapData}
\alias{LobsterHSI_mapData}
\title{Lobster HSI Map Data}
\format{
A dataframe with rows = 4066, columns = Y (Longitude, Latitude, SST, Depth, Season).
}
\usage{
LobsterHSI_mapData
}
\description{
Spatial dataset prepared for mapping lobster HSI values.
}
\keyword{datasets}
