% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LobsterHSI_monthData}
\alias{LobsterHSI_monthData}
\title{Lobster Monthly Data}
\format{
A dataframe with rows = 4066, columns = Y (Longitude, Latitude, SST, Depth, Month).
}
\usage{
LobsterHSI_monthData
}
\description{
Monthly spatial dataset prepared for mapping lobster HSI values.
}
\keyword{datasets}
