% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_SI_curves.R
\name{plot_SI_curves}
\alias{plot_SI_curves}
\title{Plot Suitability Index (SI) Curves}
\usage{
plot_SI_curves(surveydata, envVariable, Seasons, int_n, filename)
}
\description{
Plots SI curves for one or multiple environmental variables.
Currently only capable of plotting spring and fall
}
\examples{
plots=plot_SI_curves(surveydata = Lobster_SI_9020,
envVariable = c("Latitude","Longitude","Depth", "SST"),
int_n = 20,
Seasons =c("Spring","Fall"),
filename = "SI_curvrs.jpg")
}
