% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIcurve.R
\name{SIcurve}
\alias{SIcurve}
\title{Generate Suitability Index (SI) Curve}
\usage{
SIcurve(surveydata, envVariable, Seasons, int_n)
}
\value{
A dataframe of SI curve bins, SI values and SI breaks.
}
\description{
Generates SI curves for a given environmental variable.
}
\examples{
SIData=SIcurve(surveydata = Lobster_SI_9020,
envVariable = c("Latitude","Longitude","Depth", "SST"),
int_n = 20,
Seasons =c("Spring","Fall"))
}
